+++ node/src/Transaction.ts
@@ -48,6 +48,7 @@ import {
     SortClusterOptions,
     SortOptions,
     StreamAddOptions,
+    StreamGroupOptions,
     StreamReadOptions,
     StreamTrimOptions,
     ZAddOptions,
@@ -93,6 +94,7 @@ import {
     createGet,
     createGetBit,
     createGetDel,
+    createGetRange,
     createHDel,
     createHExists,
     createHGet,
@@ -186,6 +188,8 @@ import {
     createXLen,
     createXRead,
     createXTrim,
+    createXGroupCreate,
+    createXGroupDestroy,
     createZAdd,
     createZCard,
     createZCount,
@@ -286,6 +290,25 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      */
     public getdel(key: string): T {
         return this.addAndReturn(createGetDel(key));
+    }
+
+    /**
+     * Returns the substring of the string value stored at `key`, determined by the offsets
+     * `start` and `end` (both are inclusive). Negative offsets can be used in order to provide
+     * an offset starting from the end of the string. So `-1` means the last character, `-2` the
+     * penultimate and so forth. If `key` does not exist, an empty string is returned. If `start`
+     * or `end` are out of range, returns the substring within the valid range of the string.
+     *
+     * See https://valkey.io/commands/getrange/ for details.
+     *
+     * @param key - The key of the string.
+     * @param start - The starting offset.
+     * @param end - The ending offset.
+     *
+     * Command Response - substring extracted from the value stored at `key`.
+     */
+    public getrange(key: string, start: number, end: number): T {
+        return this.addAndReturn(createGetRange(key, start, end));
     }
 
     /** Set the given key with the given value. Return value is dependent on the passed options.
@@ -2120,6 +2143,44 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      */
     public xinfoConsumers(key: string, group: string): T {
         return this.addAndReturn(createXInfoConsumers(key, group));
+    }
+
+    /**
+     * Creates a new consumer group uniquely identified by `groupname` for the stream
+     * stored at `key`.
+     *
+     * See https://valkey.io/commands/xgroup-create/ for more details.
+     *
+     * @param key - The key of the stream.
+     * @param groupName - The newly created consumer group name.
+     * @param id - Stream entry ID that specifies the last delivered entry in the stream from the new
+     *     group’s perspective. The special ID `"$"` can be used to specify the last entry in the stream.
+     *
+     * Command Response - `"OK"`.
+     */
+    public xgroupCreate(
+        key: string,
+        groupName: string,
+        id: string,
+        options?: StreamGroupOptions,
+    ): T {
+        return this.addAndReturn(
+            createXGroupCreate(key, groupName, id, options),
+        );
+    }
+
+    /**
+     * Destroys the consumer group `groupname` for the stream stored at `key`.
+     *
+     * See https://valkey.io/commands/xgroup-destroy/ for more details.
+     *
+     * @param key - The key of the stream.
+     * @param groupname - The newly created consumer group name.
+     *
+     * Command Response - `true` if the consumer group is destroyed. Otherwise, `false`.
+     */
+    public xgroupDestroy(key: string, groupName: string): T {
+        return this.addAndReturn(createXGroupDestroy(key, groupName));
     }
 
     /**